% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_config}
\alias{get_config}
\title{Returns a packet with the camera's current settings.}
\usage{
get_config(con)
}
\arguments{
\item{con}{open socket connection from \code{\link[=tcam_connect]{tcam_connect()}}}
}
\description{
Returns a packet with the camera's current settings.
}
