% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_image}
\alias{get_image}
\title{Returns a packet with metadata, radiometric (or AGC) image data and Lepton telemetry objects.}
\usage{
get_image(con)
}
\arguments{
\item{con}{open socket connection from \code{\link[=tcam_connect]{tcam_connect()}}}
}
\description{
Returns a packet with metadata, radiometric (or AGC) image data and Lepton telemetry objects.
}
