% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_lep_cci}
\alias{get_lep_cci}
\title{Reads and returns specified data from the Lepton's CCI interface.}
\usage{
get_lep_cci(con, command, length)
}
\arguments{
\item{con}{open socket connection from \code{\link[=tcam_connect]{tcam_connect()}}}

\item{command}{decimal representation of the 16-bit Lepton \code{COMMAND} register
value. For example, the value \code{20172} above is \code{0x4ECC} (RAD Spotmeter
Region of Interest).}

\item{length}{decimal number of 16-bit words to read (\code{1-512}).}
}
\description{
Reads and returns specified data from the Lepton's CCI interface.
}
\references{
\href{https://www.flir.com/globalassets/imported-assets/document/flir-lepton-software-interface-description-document.pdf}{FLIR LEPTON Software Interface Description Document (IDD)}
}
