% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_status}
\alias{get_status}
\title{Returns a packet with camera status.}
\usage{
get_status(con)
}
\arguments{
\item{con}{open socket connection from \code{\link[=tcam_connect]{tcam_connect()}}}
}
\description{
Returns a packet with camera status.
}
