% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_wifi}
\alias{get_wifi}
\title{Returns a packet with the camera's current WiFi and Network configuration.}
\usage{
get_wifi(con)
}
\arguments{
\item{con}{open socket connection from \code{\link[=tcam_connect]{tcam_connect()}}}
}
\description{
Returns a packet with the camera's current WiFi and Network configuration.
}
