% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{set_time}
\alias{set_time}
\title{Set the camera's clock.}
\usage{
set_time(con, time = Sys.time())
}
\arguments{
\item{con}{open socket connection from \code{\link[=tcam_connect]{tcam_connect()}}}

\item{time}{\code{Date} or \code{POSIXct} object or a character string that
can be coerced to one; defaults to "now".}
}
\description{
Set the camera's clock.
}
