% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcam-package.R
\docType{package}
\name{tcam}
\alias{tcam}
\alias{tcam-package}
\title{Retrieve Radiometric Image Data from and Configure and Control tCam and tCam-Mini
Thermal Imaging Systems}
\description{
The tCam and tCam-Mini (\url{https://github.com/danjulio/lepton/tree/master/ESP32})
are two cameras designed around the ESP32 chipset and provide easy access to
radiometric data from Lepton 3.5 sensors. Tools are provided to configure, control,
and receive radiometric data from tCam systems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/tcam}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/tcam/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
