% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{tcam_connect}
\alias{tcam_connect}
\title{Make a socket connection to a tCam device}
\usage{
tcam_connect(host = "192.168.4.1", port = 5001)
}
\arguments{
\item{host, port}{IP/hostname + port; defaults to the device defaults}
}
\description{
Make a socket connection to a tCam device
}
