% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_radiometric}
\alias{tidy_radiometric}
\title{Return a tidy data frame of Lepton radiometric data retrieved with \code{\link[=get_image]{get_image()}}}
\usage{
tidy_radiometric(img_data)
}
\arguments{
\item{img_data}{data structure retrieved with \code{\link[=get_image]{get_image()}}}
}
\value{
data frame of \code{x}, \code{y}, and \code{value}
}
\description{
Converts the base64 16-bit coded Kelvin radiometric data to a tidy "XYZ"
data frame in Fahrenheit.
}
